Training and development
Chapter 3 focuses on the use of information technology in various
business functions. Review the business functions that Gardeners+ needs
to carry out and make a list. Go online to research an application that
will support as many of Gardeners+’s functions as possible. Explain why
you believe this application is a good choice.